+++ src/components/PlayerView/properties.ts
@@ -79,8 +79,6 @@ export interface BasePlayerViewProps {
    * When disabled, `isPictureInPictureRequested` property has no effect.
    *
    * Can be changed dynamically after the player is initialized.
-   *
-   * @platform iOS, tvOS
    */
   isPictureInPictureEnabled?: boolean;
 }